added Twitter Feed, HTML5 Template

+++ dr4k4n.py
@@ -1,22 +1,22 @@
-#!/usr/bin/env python
+﻿#!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
 from flask import Flask, request, session, g, redirect, url_for, abort, render_template, flash
 from flaskext.sqlalchemy import SQLAlchemy
 from hashlib import md5
+from twython import Twython
 
 SECRET_KEY = 'VERYS3CR3T!'
 DEBUG = True
-HOST = '127.0.0.1'
-
-ADMIN_USERNAME = 'Dr4K4n'
-ADMIN_PASSWORD = 'changeme'
+HOST = '0.0.0.0'
 
 app = Flask(__name__)
 app.config.from_object(__name__)
 app.config.from_envvar('DR4K4N_SETTINGS', silent=True)
 app.config['SQLALCHEMY_DATABASE_URI'] = 'mysql://dr4k4n:dr4k4n@localhost/dr4k4n'
 db = SQLAlchemy(app)
+
+twitter = Twython()
 
 # DB Models
 class User(db.Model):
@@ -42,8 +42,6 @@ class User(db.Model):
     password = property(_get_password,_set_password)
 
     def check_password(self, password):
-        print self.password
-        print md5(password).hexdigest()
         return self.password == md5(password).hexdigest()
     
     def is_admin(self):
@@ -82,27 +80,39 @@ class Comment(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     author_id = db.Column(db.Integer, db.ForeignKey('users.id'))
     author = db.relationship(User, uselist=False, backref=db.backref('commentUsers',cascade="none"))
+    post_id = db.Column(db.Integer, db.ForeignKey('blogposts.id'))
+    post = db.relationship(BlogPost, uselist=False, backref=db.backref('commentPost',cascade="none"))
     comment = db.Column(db.Text, nullable=False)
     date = db.Column(db.DateTime, nullable=False)
     
-    def __init__(self, author, comment, date):
-        self.author = author
-        self.comment = comment
-        self.date = date
-    
+    def __init__(self, post_id, author_id, comment):
+		self.post_id = post_id
+		self.author_id = author_id
+		self.comment = comment
+
+    def getComment(self):
+        return unicode.replace(self.comment,'\n','<br>')	
+        
     def __repr__(self):
         return '<Comment %r>' % self.id
 
-# static pages
+# main pages
 @app.route('/')
 def home():
     posts = BlogPost.query.limit(10).all()
     return render_template('main.html', current='home', posts=posts)
 
-@app.route('/read/<int:post_id>')
+@app.route('/read/<int:post_id>', methods=['GET','POST'])
 def read_post(post_id):
+    if request.form and request.form['comment'] != "":
+		comment = Comment(post_id,session['user_id'],request.form['comment'])
+		db.session.add(comment)
+		db.session.commit()
+		flash(u'Danke für deinen Kommentar')		
     post = BlogPost.query.get(post_id)
-    return render_template('read.html', current='home', post=post)
+    comments = Comment.query.filter(Comment.post_id == post.id).all()
+    logged_in = ('user_id' in session)
+    return render_template('read.html', current='home', post=post, comments=comments, logged_in=logged_in)
 
 @app.route('/gallery')
 def gallery():
@@ -117,25 +127,19 @@ def impressum():
 def login():
     error = None
     if request.method == 'POST':
-        if request.form['username'] == ADMIN_USERNAME and request.form['password'] == ADMIN_PASSWORD:
-            flash(u'Tach Chef!')
-            session['logged_in'] = True
-            session['admin'] = True
-            session['user_id'] = 1
-        else:
-            user = User.query.filter_by(username=request.form['username']).first()
-            if user and user.check_password(request.form['password']):
-                session['logged_in'] = True
-                session['user_id'] = user.id
-                session['admin'] = user.admin
-                if user.admin:
-                    flash(u'Moinsen')
-                    return redirect(url_for('home'))
-                else:
-                    flash(u'Tach Chef!')
-                    return redirect(url_for('admin'))
-            else:
-                error = u'Öhm... verkehrt :-('
+		user = User.query.filter_by(username=request.form['username']).first()
+		if user and user.check_password(request.form['password']):
+			session['logged_in'] = True
+			session['user_id'] = user.id
+			session['admin'] = user.admin
+			if user.admin:
+				flash(u'Moinsen')
+				return redirect(url_for('home'))
+			else:
+				flash(u'Tach Chef!')
+				return redirect(url_for('intern'))
+		else:
+			error = u'Öhm... verkehrt :-('
         
     return render_template('login.html', current='login', error=error)    
 
@@ -149,7 +153,7 @@ def logout():
 @app.route('/search',methods=['GET'])
 def search():
     q = request.args.get('q')
-    if len(q) > 3:
+    if len(q) > 2:
         posts = BlogPost.query.from_statement('SELECT * FROM blogposts WHERE title LIKE "%user%"').all()
     else:        
         flash('Leider nix gefunden... hast du mindestens 3 Zeichen eingegeben ?')
@@ -157,16 +161,13 @@ def search():
     return render_template('search.html', current='home', posts=posts)
 
 # admin menu 
-@app.route('/admin') 
-def admin():
+@app.route('/intern') 
+def intern():
     if not session.get('logged_in'):
         flash(u'ohne Login wird das aber nix...')
         return redirect(url_for('login'))
-    if not session.get('admin'):
-        flash(u'Hier haste nix verloren - husch husch !')        
-        return redirect(url_for('home'))
-        
-    return render_template('admin.html', current='admin')
+
+    return render_template('intern.html', current='intern', admin=session.get('admin'))
 
 # user verwaltung
 @app.route('/user/verwalten')
@@ -174,50 +175,53 @@ def user_verwalten():
     if not session.get('logged_in'):        
         flash(u'ohne Login wird das aber nix...')
         return redirect(url_for('login'))
+    if not session.get('admin'):
+        flash(u'pöse!!!')
+        return redirect(url_for('home'))
     
     users = User.query.all()
     
-    return render_template('user_verwalten.html', current='admin', users=users)
+    return render_template('user_verwalten.html', current='intern', users=users)
 
 @app.route('/user/new',methods=['GET', 'POST'])
 def user_new():
     if not session.get('logged_in'):        
         flash(u'ohne Login wird das aber nix...')
         return redirect(url_for('login'))
-    
+    if not session.get('admin'):
+        flash(u'pöse!!!')
+        return redirect(url_for('home'))
+        
     if not request.form:
-        return render_template('user_new.html', current='admin');    
-    
-    print '1'
+        return render_template('user_new.html', current='intern');    
     
     newUserAdmin = 0
     if 'admin' in request.form and request.form['admin'] == 'on':
         newUserAdmin = 1
-    
-    print '2'
-     
+
     newUser = User(request.form['username'], request.form['email'], request.form['password'], newUserAdmin)
-    
-    print '3'
-    
+
     db.session.add(newUser)
     db.session.commit()
     
     flash(u'nice one - is drin ;-)')
-    return redirect(url_for('admin'))
+    return redirect(url_for('intern'))
 
 @app.route('/user/<int:user_id>/edit',methods=['GET','POST'])
 def user_edit(user_id):
     if not session.get('logged_in'):        
         flash(u'ohne Login wird das aber nix...')
         return redirect(url_for('login'))
-        
+    if not session.get('admin'):
+        flash(u'pöse!!!')
+        return redirect(url_for('home'))
+                
     user = User.query.get(user_id)
     newUser = user
     
     if user is None:
         flash(u'Bidde? den kennsch ned!')
-        return redirect(url_for('admin'))
+        return redirect(url_for('intern'))
     
     if request.form:
         editUserAdmin = 0
@@ -227,30 +231,33 @@ def user_edit(user_id):
             user.password = request.form['password']
         user.username = request.form['username']
         user.email = request.form['email']
-        user.admin = newUserAdmin
+        user.admin = editUserAdmin
         db.session.commit()
         flash(u'check! Hab ich')
-        return redirect(url_for('admin'))
+        return redirect(url_for('intern'))
         
-    return render_template('user_edit.html', current='admin',user=user);
+    return render_template('user_edit.html', current='intern',user=user);
 
 @app.route('/user/<int:user_id>/del',methods=['GET'])
 def user_del(user_id):
     if not session.get('logged_in'):        
         flash(u'ohne Login wird das aber nix...')
         return redirect(url_for('login'))
-        
+    if not session.get('admin'):
+        flash(u'pöse!!!')
+        return redirect(url_for('home'))
+                
     user = User.query.get(user_id)
     
     if user is None:
         flash(u'Bidde? den kennsch ned!')
-        return redirect(url_for('admin'))
+        return redirect(url_for('intern'))
     
     db.session.delete(user)
     db.session.commit()
     
     flash(u'Atze entfernt')
-    return redirect(url_for('admin'))
+    return redirect(url_for('intern'))
 
 # blog verwaltung
 @app.route('/bloggen/verwalten')
@@ -261,7 +268,7 @@ def bloggen_verwalten():
     
     blogPosts = BlogPost.query.all()
     
-    return render_template('bloggen_verwalten.html', current='admin', blogPosts=blogPosts)
+    return render_template('bloggen_verwalten.html', current='intern', blogPosts=blogPosts)
 
 @app.route('/bloggen/new',methods=['GET', 'POST'])
 def bloggen_new():
@@ -270,13 +277,13 @@ def bloggen_new():
         return redirect(url_for('login'))
     
     if not request.form:
-        return render_template('bloggen_new.html', current='admin');
+        return render_template('bloggen_new.html', current='intern');
     
     newPost = BlogPost(request.form['title'], session['user_id'], request.form['shorttext'], request.form['longtext'])
     db.session.add(newPost)
     db.session.commit()
     flash(u'nice one - is drin ;-)')
-    return redirect(url_for('admin'))
+    return redirect(url_for('intern'))
  
 @app.route('/bloggen/<int:post_id>/edit',methods=['GET', 'POST'])
 def bloggen_edit(post_id):
@@ -288,7 +295,7 @@ def bloggen_edit(post_id):
 
     if post is None:
         flash(u'Bidde? den kennsch ned!')
-        return redirect(url_for('admin'))
+        return redirect(url_for('intern'))
     
     if request.form:
         post.title = request.form['title']
@@ -296,9 +303,9 @@ def bloggen_edit(post_id):
         post.longtext = request.form['longtext']
         db.session.commit()
         flash(u'check! Hab ich')
-        return redirect(url_for('admin'))  
+        return redirect(url_for('intern'))  
     
-    return render_template('bloggen_edit.html', current='admin', post=post);
+    return render_template('bloggen_edit.html', current='intern', post=post);
 
 @app.route('/bloggen/<int:post_id>/del',methods=['GET', 'POST'])
 def bloggen_del(post_id):
@@ -310,14 +317,26 @@ def bloggen_del(post_id):
     
     if post is None:
         flash(u'Bidde? den kennsch ned!')
-        return redirect(url_for('admin'))
+        return redirect(url_for('intern'))
     
     db.session.delete(post)
     db.session.commit()
     
     flash(u'Post entfernt')
-    return redirect(url_for('admin'))
+    return redirect(url_for('intern'))
 
+@app.route('/backend/getTweets')
+def get_tweets():
+	tweetsAll = twitter.getUserTimeline(screen_name='Dr4K4n')
+	tweets = []
+	t_i = 0
+	for t in tweetsAll:
+		if t_i > 9:
+			break
+		tweets.append([t['id'],t['text']])
+		t_i += 1
+	return render_template('backend/tweets.html', tweets=tweets)
+	
 # start
 if __name__ == '__main__':
     app.run(debug=app.config['DEBUG'],host=app.config['HOST'])
